+++ CHANGELOG.md
@@ -8,6 +8,7 @@ LUYA CHANGELOG
 
 **ISSUES**
 
+- `#726` Crawler module does now ignore not text based response header.
 - `#308` Added copy and paste for cms blocks
 - `#716` Add is_deleted files to the storag container list.
 - `#715` CMS nav item model does now slugify the alias in the model before create/update to ensure correct alias paths.